Draft: Make sure all triggers with "*" check for directories

See merge request alpine/aports!57032

+++ main/gtk+3.0/gtk-update-icon-cache.trigger
@@ -1,9 +1,8 @@
 #!/bin/sh
 
 for i in "$@"; do
-	if ! [ -e "$i" ]; then
-		continue
-	fi
+	[ -e "$i" ] || continue
+	[ -d "$i" ] || continue
 	gtk-update-icon-cache -q -t -f "$i"
 	rmdir "$1" 2>/dev/null || :
 done
